+++ libs/base/src/code/result/mod.rs
@@ -3,6 +3,7 @@
 #![doc = crate::_DOC_CODE_RESULT!()]
 //
 
+mod mismatch; // Mismatch
 mod reexports;
 
 crate::items! { // structural access: _mods, _all
@@ -10,7 +11,7 @@ crate::items! { // structural access: _mods, _all
     pub use _mods::*;
 
     mod _mods {
-        pub use super::reexports::*;
+        pub use super::{mismatch::*, reexports::*};
     }
     pub(super) mod _all {
         #[doc(inline)]