+++ libs/base/src/data/mod.rs
@@ -3,9 +3,11 @@
 #![doc = crate::_DOC_DATA!()]
 //
 
-mod sort;
+// mod bit; // Bitwise, bitfield!
+mod sort; // Sort
 
 pub mod codec;
+pub mod errors;
 pub mod iter;
 pub mod list;
 // pub mod key;
@@ -15,11 +17,13 @@ crate::items! { // structural access: _mods, _pub_mods, _all
     pub use _mods::*;
 
     mod _mods { #![allow(unused)]
+        // pub use super::bit::_all::*; // TODO
         pub use super::sort::_all::*;
     }
     mod _pub_mods {
         pub use super::{
             codec::_all::*,
+            errors::*,
             iter::_all::*,
             list::_all::*,
             // key::_all::*,