+++ libs/base_std/src/phys/time/errors.rs
@@ -1,27 +1,21 @@
-// devela::code::error::time
+// devela_base_std::phys::time::errors
 //
 //! Time-related errors.
 //
 // TOC
-// - individual data-related error types:
+// - individual errors:
 //   - SystemTimeError
-//   - Timeout
 // - partial composite errors:
 // - full composite errors:
 //   - TimeError
 //   - TimeResult
 
-use crate::define_error;
-#[cfg(feature = "std")]
-use {
-    crate::{Duration, MpscRecvTimeoutError},
-    ::std::time::SystemTimeError as StdSystemTimeError,
-};
+use crate::{Duration, define_error, StdSystemTimeError};
+// use ::std::time::SystemTimeError as StdSystemTimeError;
 
 /* individual errors */
 
 define_error! { individual:
-    #[cfg(feature = "std")]
     #[cfg_attr(nightly_doc, doc(cfg(feature = "std")))]
     pub struct SystemTimeError(Duration);
     +tag: crate::TAG_TIME!(),
@@ -30,29 +24,30 @@ define_error! { individual:
 This is basically a replication of `std::time::`[`SystemTimeError`][StdSystemTimeError].",
     self+f => write!(f, "SystemTimeError difference: {:?}", self.0)
 }
-#[cfg(feature = "std")] #[rustfmt::skip]
-#[cfg_attr(nightly_doc, doc(cfg(feature = "std")))]
+
+#[rustfmt::skip]
+#[cfg_attr(nightly_doc, doc(cfg(feature = "std")))] // RETHINK these
 impl From<StdSystemTimeError> for SystemTimeError {
     fn from(from: StdSystemTimeError) -> Self { SystemTimeError(from.duration()) }
 }
 
-define_error! { individual: pub struct Timeout;
-    DOC_KEY_ALREADY_EXISTS = "The operation has exceeded the allowed execution time.",
-    self+f => write!(f, "The operation has exceeded the allowed execution time.")
-}
-#[cfg(feature = "std")] #[rustfmt::skip]
-#[cfg_attr(nightly_doc, doc(cfg(feature = "std")))]
-impl From<MpscRecvTimeoutError> for Timeout {
-    fn from(_from: MpscRecvTimeoutError) -> Self { Timeout }
-}
+// NOTE: can't implement this conversion here because it needs std…
+// but Timeout is maybe important enough to be defined here.
+// MAYBE we could do alternative versions for no_std and std.
+//
+// use crate::{MpscRecvTimeoutError, Timeout};
+// impl From<MpscRecvTimeoutError> for Timeout {
+//     fn from(_from: MpscRecvTimeoutError) -> Self { Timeout }
+// }
 
-#[cfg(all(feature = "error", feature = "time"))]
 pub use full_composite::*;
-#[cfg(all(feature = "error", feature = "time"))]
-#[cfg_attr(nightly_doc, doc(cfg(all(feature = "error", feature = "time"))))]
 mod full_composite {
     use super::*;
-    use crate::{DOC_DATA_OVERFLOW, DataOverflow};
+    use crate::{CONST, DataOverflow};
+
+    CONST! {
+        DOC_DATA_OVERFLOW = "The value has surpassed the bounds of the representable data space.";
+    }
 
     define_error! { composite: fmt(f)
         #[doc = crate::TAG_TIME!()]
@@ -62,22 +57,13 @@ mod full_composite {
             DOC_DATA_OVERFLOW: +const
                 DataOverflow(o|0: Option<usize>) => DataOverflow(*o),
 
-            #[cfg(feature = "std")]
-            #[cfg_attr(nightly_doc, doc(cfg(feature = "std")))]
             DOC_SYSTEM_TIME_ERROR: +const
                 SystemTime(d|0: Duration) => SystemTimeError(*d),
         }
     }
-    #[cfg(feature = "std")]
-    #[cfg_attr(nightly_doc, doc(cfg(feature = "std")))]
     impl From<StdSystemTimeError> for TimeError {
         fn from(from: StdSystemTimeError) -> Self {
             TimeError::SystemTime(from.duration())
         }
     }
-
-    #[doc = crate::TAG_TIME!()]
-    #[doc = crate::TAG_RESULT!()]
-    /// A time-related result.
-    pub type TimeResult<T> = crate::Result<T, TimeError>;
 }