+++ libs/base_std/src/phys/time/reexports.rs
@@ -3,7 +3,7 @@
 //!
 //
 
-use crate::{_reexport, TAG_TIME};
+use crate::{_reexport, TAG_ERROR, TAG_TIME};
 
 _reexport! { rust: std::time,
     tag: TAG_TIME!(),
@@ -15,12 +15,12 @@ _reexport! { rust: std::time,
     doc: "A measurement of the system clock.",
     SystemTime
 }
-// _reexport! { rust: std::time,
-//    tag: concat![TAG_TIME!(), TAG_ERROR!()],
-//     tag: crate::TAG_ERROR!(),
-//     doc: "Error returned from the `duration_since` and `elapsed` methods on [`SystemTime`].",
-//     SystemTimeError
-// }
+// NOTE: replicated in phys::time:errors
+_reexport! { rust: std::time,
+    tag: concat![TAG_TIME!(), TAG_ERROR!()],
+    doc: "Error returned from the `duration_since` and `elapsed` methods on [`SystemTime`].",
+    @SystemTimeError as StdSystemTimeError
+}
 _reexport! { rust: std::time,
     tag: TAG_TIME!(),
     doc: "A [`SystemTime`] anchored to “1970-01-01 00:00:00 UTC”.",