+++ src/code/result/mod.rs
@@ -17,12 +17,10 @@
 //! and mismatch-aware comparisons ([`Mismatch`]).
 //
 
-crate::mod_path!(_c "../../../libs/base/src/code/result/reexports.rs");
-
 mod chain_hook; // Chain, Hook
-mod mismatch; // Mismatch
 mod opt_res; // ExtOption, ExtResult, OptRes, ExtOptRes, sok, serr, OptionFmt[Or[Else]]
 mod own; // Own
+mod reexports;
 mod value_quant; // ValueQuant
 
 crate::items! { // structural access: _mods, _all, _always
@@ -33,7 +31,7 @@ crate::items! { // structural access: _mods, _all, _always
 
     mod _mods {
         pub use super::{
-            _c::*, chain_hook::*, mismatch::*, opt_res::_all::*, own::*, value_quant::*,
+            chain_hook::*, opt_res::_all::*, own::*, reexports::*, value_quant::*,
         };
         // WIPZONE
         // pub use super::enumatch::*;
@@ -45,7 +43,7 @@ crate::items! { // structural access: _mods, _all, _always
         pub use super::_mods::*;
     }
     pub(super) mod _always { #![allow(unused)]
-        pub use super::_c::*;
+        pub use super::reexports::*;
     }
 }
 // WIPZONE